add section roles,changed menu at sleepingowl

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,7 +20,7 @@ class CreateUsersTable extends Migration
             $table->string('password');
             $table->integer('role_id');
 	        $table->string('image');
-	        $table->integer('balance');
+	        $table->double('balance');
 	        $table->rememberToken();
             $table->timestamps();
         });
